add logout
