+++ atoms/atom.ts
@@ -1,3 +1,4 @@
 import { atom } from 'recoil'
 
 export const LoginModal = atom<boolean>({ key: 'LoginModal', default: false })
+export const HasLogin = atom<boolean>({ key: 'HasLogin', default: false })
