+++ types/LayoutType.ts
@@ -1,0 +1,3 @@
+export type LayoutType = {
+  children: React.ReactNode
+}
